Bug fix
The percentage counter in the scoresheet section of the code in main.py was hardcoded, which resulted in false results. The way the user enters vocabulary words was also changed. The GUI program was also deleted as it was not fulfuil the program's function. A better GUI may be added latter.